Add recoveries datapoints
Bumped versions of puppeteer and inquiry
Added imported value. Updated total to be a calculated value. Updated numeric datapoints to return as numbers in JSON.

Breaking changes, output format is now nested
This changes both the JSON properties and the headers.
Chose to make this breaking change as fixing the addition of the Recoveries breakout was in itself breaking.